Move build run result related code into build_run_result.go
