+++ models/models.go
@@ -207,32 +207,6 @@ type BuildRunStartModel struct {
 	StartTime   time.Time `json:"start_time" yaml:"start_time"`
 }
 
-type BuildRunResultsModel struct {
-	WorkflowID           string                `json:"workflow_id" yaml:"workflow_id"`
-	EventName            string                `json:"event_name" yaml:"event_name"`
-	ProjectType          string                `json:"project_type" yaml:"project_type"`
-	StartTime            time.Time             `json:"start_time" yaml:"start_time"`
-	StepmanUpdates       map[string]int        `json:"stepman_updates" yaml:"stepman_updates"`
-	SuccessSteps         []StepRunResultsModel `json:"success_steps" yaml:"success_steps"`
-	FailedSteps          []StepRunResultsModel `json:"failed_steps" yaml:"failed_steps"`
-	FailedSkippableSteps []StepRunResultsModel `json:"failed_skippable_steps" yaml:"failed_skippable_steps"`
-	SkippedSteps         []StepRunResultsModel `json:"skipped_steps" yaml:"skipped_steps"`
-}
-
-type StepRunResultsModel struct {
-	StepInfo   stepmanModels.StepInfoModel `json:"step_info" yaml:"step_info"`
-	StepInputs map[string]string           `json:"step_inputs" yaml:"step_inputs"`
-	Status     StepRunStatus               `json:"status" yaml:"status"`
-	Idx        int                         `json:"idx" yaml:"idx"`
-	RunTime    time.Duration               `json:"run_time" yaml:"run_time"`
-	StartTime  time.Time                   `json:"start_time" yaml:"start_time"`
-	ErrorStr   string                      `json:"error_str" yaml:"error_str"`
-	ExitCode   int                         `json:"exit_code" yaml:"exit_code"`
-
-	Timeout         time.Duration `json:"-"`
-	NoOutputTimeout time.Duration `json:"-"`
-}
-
 type StepError struct {
 	Code    int    `json:"code"`
 	Message string `json:"message"`
